starting a ray tracing values server for ionization param x-ray treatment and pdr-calculations
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@4159 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/pdr_mod.F90
@@ -422,7 +422,7 @@ END SUBROUTINE CALCULATE_DUST_TEMPERATURES
 recursive subroutine iniTempGuess(thisOctal)
   integer :: subcell, i
   type(octal), pointer :: thisOctal, child
-  real :: tguess
+  real(double) :: tguess
   real(double), parameter :: tmin = 10.d0
   real(double), parameter :: tmax = 1.d4
